Move queue into own namespace
Moved queue-related interfaces, implementations, and models from Harvester and Core projects into new Fip.Strive.Queue.Application and Fip.Strive.Queue.Domain projects.

+++ src/harvest/Fip.Strive.Harvester.Application/Features/Import/Services/ImportWorker.cs
@@ -1,12 +1,11 @@
-using Fip.Strive.Core.Domain.Schemas.Queue.Enums;
-using Fip.Strive.Core.Domain.Schemas.Queue.Models;
-using Fip.Strive.Core.Domain.Schemas.Queue.Models.Signals;
-using Fip.Strive.Harvester.Application.Core.Queue.Components;
 using Fip.Strive.Harvester.Application.Features.Import.Services.Contracts;
+using Fip.Strive.Harvester.Domain.Signals;
+using Fip.Strive.Queue.Application.Components;
+using Fip.Strive.Queue.Domain.Models;
 
 namespace Fip.Strive.Harvester.Application.Features.Import.Services;
 
-public class ImportWorker(IImportService inventory) : QueueWorker(SignalTypes.UploadSignal)
+public class ImportWorker(IImportService inventory) : QueueWorker((int)SignalTypes.UploadSignal)
 {
     public override async Task ProcessAsync(JobDetails job, CancellationToken ct)
     {
